update for mach-core -> mach repository change
Mach core package is now available via `@import("mach").core` in the main
repository instead of as a seperate repository. We still have a few issues
to sort out, but this approach ultimately will just mean the code moved
locations. Because Zig has lazy code evaluation and lazy dependency support
in the latest version (we're not using it yet, but will soon) this will
effectively mean no difference to you as a user in terms of build time or
dependencies that need to be downloaded.

See hexops/mach#1165 for motivation behind this change

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>